+++ test/perf/perf.cc
@@ -1,0 +1,50 @@
+/*
+ * Copyright (C) 2021 ScyllaDB
+ */
+
+/*
+ * This file is part of Scylla.
+ *
+ * Scylla is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Scylla is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Scylla.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "perf.hh"
+#include "seastarx.hh"
+
+void scheduling_latency_measurer::schedule_tick() {
+    seastar::schedule(make_task(default_scheduling_group(), [self = weak_from_this()] () mutable {
+        if (self) {
+            self->tick();
+        }
+    }));
+}
+
+std::ostream& operator<<(std::ostream& out, const scheduling_latency_measurer& slm) {
+    auto to_ms = [] (int64_t nanos) {
+        return float(nanos) / 1e6;
+    };
+    return out << sprint("{count: %d, "
+                         //"min: %.6f [ms], "
+                         //"50%%: %.6f [ms], "
+                         //"90%%: %.6f [ms], "
+                         "99%%: %.6f [ms], "
+                         "max: %.6f [ms]}",
+        slm.histogram().count(),
+        //to_ms(slm.min().count()),
+        //to_ms(slm.histogram().percentile(0.5)),
+        //to_ms(slm.histogram().percentile(0.9)),
+        to_ms(slm.histogram().percentile(0.99)),
+        to_ms(slm.max().count()));
+}
+
