+++ test/perf/perf.hh
@@ -172,28 +172,4 @@ public:
     clk::duration max() const { return _minmax.max(); }
 };
 
-void scheduling_latency_measurer::schedule_tick() {
-    seastar::schedule(make_task(default_scheduling_group(), [self = weak_from_this()] () mutable {
-        if (self) {
-            self->tick();
-        }
-    }));
-}
-
-std::ostream& operator<<(std::ostream& out, const scheduling_latency_measurer& slm) {
-    auto to_ms = [] (int64_t nanos) {
-        return float(nanos) / 1e6;
-    };
-    return out << sprint("{count: %d, "
-                         //"min: %.6f [ms], "
-                         //"50%%: %.6f [ms], "
-                         //"90%%: %.6f [ms], "
-                         "99%%: %.6f [ms], "
-                         "max: %.6f [ms]}",
-        slm.histogram().count(),
-        //to_ms(slm.min().count()),
-        //to_ms(slm.histogram().percentile(0.5)),
-        //to_ms(slm.histogram().percentile(0.9)),
-        to_ms(slm.histogram().percentile(0.99)),
-        to_ms(slm.max().count()));
-}
+std::ostream& operator<<(std::ostream& out, const scheduling_latency_measurer& slm);
